apply fixes discovered during testing for host IP discovery, logging, and message delineation
It turns out that there is no way, using standard library functions, to discover
the IP addresses available at a host's network interfaces because this is not a
portable functionality.  The assignment specifies that the program need only run
on UNIX systems, in which case calling into ioctl to find these host's current
IP addresses should be portable enough.  This required a new dependency on the
extunix library and some minor modification and addition of logging calls to
`establish_chat`.  The expected behavior is now that the server starts and
displays all of its non-loopback IP addresses and listening port on stderr.
For the sake of local testing, one can simply observe this log in order to find
out what IPv4 address to supply via the `-a` flag when starting the client process.

Some additional debug level logs were added; info-level logs are only sent to stderr
as needed to demonstrate that messages were registered and that acknowledgements
were received.

I also discovered that for whatever reason, the formatter I could create with
`Format.formatter_of_out_channel` didn't seem to cause any printing to the file
handle managed by the out channel.  Nonetheless, messages were properly being
printed when using `Format.err_formatter`.  Since users are free to redirect
stderr to a log file (and indeed, this is the intended usage, otherwise
inputs and outputs of the program would get interleaved) it seemed wiser simply
to use this formatter.  Separately, I decided that since this is a chat program
it would be simpler to use `In_channel.input_line` to take in messages; just because
no encoding is assumed on the receiving end, doesn't mean we can't assume a text
encoding on the receiving end.  This seems in line with the spirit of the assignment.