correct issue with role-labeling of message ids and regenerate sample logs
I had failed to distinguish between the two kinds of incoming messages and
their origins: acks have the same ID in the logs as the message they are
acknowledging.  So if an ack is received by the server, it should have the
ID of a message that originated on the server and its name should therefore
begin with "S" rather than "C", and vice versa.  If a chat is received by
the server, rather than an ack, then that message's name should start with
"C" rather than "S" because it identifies a message that originated on the
client, and vice versa.  Prior to this commit, messages received by the
server/client were named starting with "S"/"C" respectively whether they
were an ack or a chat.